requirements.txt
browser-use>=0.1.0
langchain-openai==0.2.6
python-dotenv
pandas
requests
beautifulsoup4
langchain==0.3.7
pydantic==2.11.7
# Monitoring agent dependencies
openai>=1.0.0
aiohttp>=3.8.0
asyncio